Prep for websocket
