+++ frontend/app/routes/PyramidLocalGame.tsx
@@ -1,5 +1,4 @@
 import {
-  Badge,
   Box,
   Button,
   Container,
@@ -8,243 +7,21 @@ import {
   Stack,
   Text,
 } from "@chakra-ui/react";
-import { useEffect, useMemo, useState } from "react";
+import { useMemo, useState } from "react";
 
-import { checkQuestion, getQuestion, type QuestionResponse } from "api/api";
 import { TimerLine } from "components/General/TimerLine";
 import StandardQuestionModal from "../../components/StandardQuestionModal";
 import YesNoQuestionModal from "../../components/YesNoQuestionModal";
 import SwitchAnswerModal from "components/SwitchAnswerModal";
 import GameOverModal from "../../components/GameOverModal";
 import ConcedeConfirmModal from "../../components/ConcedeConfirmModal";
-
-type Player = "player1" | "player2";
-type TileState = "neutral" | "black" | Player;
-type QuestionType = "standard" | "yes_no";
-type Phase = "idle" | "answering" | "stealing" | "switch" | "revealing";
-type GameState = "playing" | "ended";
-
-interface TileCell {
-  id: string;
-  label: string;
-  row: number;
-  col: number;
-  state: TileState;
-}
-
-interface ActiveChallenge {
-  row: number;
-  col: number;
-  ownerPlayer: Player;
-  answerPlayer: Player;
-  questionType: QuestionType;
-  question: QuestionResponse;
-}
-
-interface GameResult {
-  winner: Player;
-  message: string;
-}
-
-const BOARD_SIZE = 7;
-const PICK_SECONDS = 10;
-const ANSWER_SECONDS = 18;
-const ANSWER_REVEAL_MS = 2200;
-
-const PLAYER_META: Record<
-  Player,
-  { label: string; color: string; softColor: string; sideGoal: string }
-> = {
-  player1: {
-    label: "Player 1",
-    color: "blue.500",
-    softColor: "blue.50",
-    sideGoal: "Connect left, right, and bottom",
-  },
-  player2: {
-    label: "Player 2",
-    color: "orange.500",
-    softColor: "orange.50",
-    sideGoal: "Connect left, right, and bottom",
-  },
-};
-
-function otherPlayer(player: Player): Player {
-  return player === "player1" ? "player2" : "player1";
-}
-
-function createBoard(): TileCell[][] {
-  return Array.from({ length: BOARD_SIZE }, (_, row) =>
-    Array.from({ length: row + 1 }, (_, col) => {
-      const index = (row * (row + 1)) / 2 + col;
-      return {
-        id: `tile-${row}-${col}`,
-        label: String(index + 1),
-        row,
-        col,
-        state: "neutral" as TileState,
-      };
-    }),
-  );
-}
-
-function isClaimed(tile: TileCell): boolean {
-  return tile.state === "player1" || tile.state === "player2";
-}
-
-function isOnSide(
-  row: number,
-  col: number,
-  side: "left" | "right" | "bottom",
-): boolean {
-  if (side === "bottom") {
-    return row === BOARD_SIZE - 1;
-  }
-
-  if (side === "left") {
-    return col === 0;
-  }
-
-  return col === row;
-}
-
-function getNeighbors(row: number, col: number): Array<[number, number]> {
-  return [
-    [row, col - 1],
-    [row, col + 1],
-    [row - 1, col - 1],
-    [row - 1, col],
-    [row + 1, col],
-    [row + 1, col + 1],
-  ];
-}
-
-function touchesAllSides(board: TileCell[][], player: Player): boolean {
-  const visited = new Set<string>();
-
-  for (let row = 0; row < board.length; row += 1) {
-    for (let col = 0; col < board[row].length; col += 1) {
-      const startTile = board[row][col];
-      const startKey = `${row}-${col}`;
-
-      if (startTile.state !== player || visited.has(startKey)) {
-        continue;
-      }
-
-      const stack: Array<[number, number]> = [[row, col]];
-      let touchesLeft = false;
-      let touchesRight = false;
-      let touchesBottom = false;
-
-      while (stack.length > 0) {
-        const [currentRow, currentCol] = stack.pop() as [number, number];
-        const key = `${currentRow}-${currentCol}`;
-
-        if (visited.has(key)) {
-          continue;
-        }
-
-        const tile = board[currentRow]?.[currentCol];
-        if (!tile || tile.state !== player) {
-          continue;
-        }
-
-        visited.add(key);
-
-        touchesLeft = touchesLeft || isOnSide(currentRow, currentCol, "left");
-        touchesRight =
-          touchesRight || isOnSide(currentRow, currentCol, "right");
-        touchesBottom =
-          touchesBottom || isOnSide(currentRow, currentCol, "bottom");
-
-        if (touchesLeft && touchesRight && touchesBottom) {
-          return true;
-        }
-
-        for (const [nextRow, nextCol] of getNeighbors(currentRow, currentCol)) {
-          const nextTile = board[nextRow]?.[nextCol];
-          if (!nextTile || nextTile.state !== player) {
-            continue;
-          }
-
-          const nextKey = `${nextRow}-${nextCol}`;
-          if (!visited.has(nextKey)) {
-            stack.push([nextRow, nextCol]);
-          }
-        }
-      }
-    }
-  }
-
-  return false;
-}
-
-function hasPotentialThreeSideConnection(
-  board: TileCell[][],
-  player: Player,
-): boolean {
-  const blockedBy = otherPlayer(player);
-  const visited = new Set<string>();
-
-  for (let row = 0; row < board.length; row += 1) {
-    for (let col = 0; col < board[row].length; col += 1) {
-      const startTile = board[row][col];
-      const startKey = `${row}-${col}`;
-
-      if (startTile.state === blockedBy || visited.has(startKey)) {
-        continue;
-      }
-
-      const stack: Array<[number, number]> = [[row, col]];
-      let touchesLeft = false;
-      let touchesRight = false;
-      let touchesBottom = false;
-
-      while (stack.length > 0) {
-        const [currentRow, currentCol] = stack.pop() as [number, number];
-        const key = `${currentRow}-${currentCol}`;
-
-        if (visited.has(key)) {
-          continue;
-        }
-
-        const tile = board[currentRow]?.[currentCol];
-        if (!tile || tile.state === blockedBy) {
-          continue;
-        }
-
-        visited.add(key);
-        touchesLeft = touchesLeft || isOnSide(currentRow, currentCol, "left");
-        touchesRight =
-          touchesRight || isOnSide(currentRow, currentCol, "right");
-        touchesBottom =
-          touchesBottom || isOnSide(currentRow, currentCol, "bottom");
-
-        if (touchesLeft && touchesRight && touchesBottom) {
-          return true;
-        }
-
-        for (const [nextRow, nextCol] of getNeighbors(currentRow, currentCol)) {
-          const nextTile = board[nextRow]?.[nextCol];
-          if (!nextTile || nextTile.state === blockedBy) {
-            continue;
-          }
-
-          const nextKey = `${nextRow}-${nextCol}`;
-          if (!visited.has(nextKey)) {
-            stack.push([nextRow, nextCol]);
-          }
-        }
-      }
-    }
-  }
-
-  return false;
-}
-
-function questionTypeForTile(tile: TileCell): QuestionType {
-  return tile.state === "black" ? "yes_no" : "standard";
-}
+import { PLAYER_META, type TileCell } from "app/game/pyramidTypes";
+import { isClaimed, otherPlayer } from "app/game/pyramidRules";
+import {
+  ANSWER_SECONDS,
+  PICK_SECONDS,
+  usePyramidGameController,
+} from "app/game/usePyramidGameController";
 
 export function meta() {
   return [
@@ -257,332 +34,26 @@ export function meta() {
 }
 
 export default function PyramidLocalGame() {
-  const [board, setBoard] = useState<TileCell[][]>(() => createBoard());
-  const [turnPlayer, setTurnPlayer] = useState<Player>(() =>
-    Math.random() < 0.5 ? "player1" : "player2",
-  );
-  const [winner, setWinner] = useState<Player | null>(null);
-  const [gameState, setGameState] = useState<GameState>("playing");
-  const [gameResult, setGameResult] = useState<GameResult | null>(null);
-  const [phase, setPhase] = useState<Phase>("idle");
-  const [pickSeconds, setPickSeconds] = useState<number>(PICK_SECONDS);
-  const [questionSeconds, setQuestionSeconds] =
-    useState<number>(ANSWER_SECONDS);
-  const [switchSeconds, setSwitchSeconds] = useState<number>(ANSWER_SECONDS);
-  const [statusPopup, setStatusPopup] = useState<string>("");
   const [showConcedeConfirm, setShowConcedeConfirm] = useState(false);
-  const [activeChallenge, setActiveChallenge] =
-    useState<ActiveChallenge | null>(null);
-
-  const startNewGame = () => {
-    const openingPlayer = Math.random() < 0.5 ? "player1" : "player2";
-
-    setBoard(createBoard());
-    setTurnPlayer(openingPlayer);
-    setWinner(null);
-    setGameState("playing");
-    setGameResult(null);
-    setPhase("idle");
-    setPickSeconds(PICK_SECONDS);
-    setQuestionSeconds(ANSWER_SECONDS);
-    setSwitchSeconds(ANSWER_SECONDS);
-    setStatusPopup("");
-    setShowConcedeConfirm(false);
-    setActiveChallenge(null);
-  };
-
-  useEffect(() => {
-    if (!winner || gameState !== "playing") {
-      return;
-    }
-
-    const timer = window.setTimeout(() => {
-      setGameState("ended");
-    }, 2500);
-
-    return () => window.clearTimeout(timer);
-  }, [winner, gameState]);
-
-  useEffect(() => {
-    if (winner || gameState !== "playing" || phase !== "idle") {
-      return;
-    }
-
-    if (pickSeconds <= 0) {
-      const winnerPlayer = otherPlayer(turnPlayer);
-      endGame(
-        winnerPlayer,
-        `${PLAYER_META[turnPlayer].label} did not pick in time. ${PLAYER_META[winnerPlayer].label} wins.`,
-      );
-      setStatusPopup(
-        `${PLAYER_META[turnPlayer].label} did not pick in time. ${PLAYER_META[winnerPlayer].label} wins.`,
-      );
-      return;
-    }
-
-    const timer = window.setTimeout(() => {
-      setPickSeconds((current) => Math.max(current - 1, 0));
-    }, 1000);
-
-    return () => window.clearTimeout(timer);
-  }, [pickSeconds, phase, turnPlayer, winner, gameState]);
-
-  useEffect(() => {
-    if (
-      gameState !== "playing" ||
-      (phase !== "answering" && phase !== "stealing" && phase !== "switch")
-    ) {
-      return;
-    }
-
-    if (phase === "switch") {
-      if (switchSeconds <= 0) {
-        void revealCorrectAndContinue();
-        return;
-      }
-
-      const switchTimer = window.setTimeout(() => {
-        setSwitchSeconds((current) => Math.max(current - 1, 0));
-      }, 1000);
-
-      return () => window.clearTimeout(switchTimer);
-    }
-
-    if (questionSeconds <= 0) {
-      void handleTimeout();
-      return;
-    }
-
-    const timer = window.setTimeout(() => {
-      setQuestionSeconds((current) => Math.max(current - 1, 0));
-    }, 1000);
-
-    return () => window.clearTimeout(timer);
-  }, [phase, questionSeconds, switchSeconds, gameState]);
-
-  const startSelection = (row: number, col: number, tile: TileCell) => {
-    if (winner || gameState !== "playing" || phase !== "idle") {
-      return;
-    }
-
-    if (isClaimed(tile)) {
-      return;
-    }
-
-    setPickSeconds(PICK_SECONDS);
-    setStatusPopup("");
-    void loadQuestion(row, col, turnPlayer, questionTypeForTile(tile));
-  };
-
-  const updateTile = (row: number, col: number, state: TileState) => {
-    setBoard((currentBoard) =>
-      currentBoard.map((currentRow, rowIndex) =>
-        rowIndex !== row
-          ? currentRow
-          : currentRow.map((tile, colIndex) =>
-              colIndex !== col ? tile : { ...tile, state },
-            ),
-      ),
-    );
-  };
-
-  const claimTile = (row: number, col: number, player: Player) => {
-    setBoard((currentBoard) => {
-      const nextBoard = currentBoard.map((currentRow, rowIndex) =>
-        rowIndex !== row
-          ? currentRow
-          : currentRow.map((tile, colIndex) =>
-              colIndex !== col ? tile : { ...tile, state: player },
-            ),
-      );
-
-      if (touchesAllSides(nextBoard, player)) {
-        endGame(
-          player,
-          `${PLAYER_META[player].label} connected all three sides and wins.`,
-        );
-        return nextBoard;
-      }
-
-      const blockedPlayer = otherPlayer(player);
-      if (!hasPotentialThreeSideConnection(nextBoard, blockedPlayer)) {
-        endGame(
-          player,
-          `${PLAYER_META[blockedPlayer].label} can no longer connect all three sides. ${PLAYER_META[player].label} wins.`,
-        );
-      }
-
-      return nextBoard;
-    });
-  };
-
-  const endGame = (winnerPlayer: Player, message: string) => {
-    setWinner(winnerPlayer);
-    setGameResult({ winner: winnerPlayer, message });
-    setGameState("ended");
-    setPhase("idle");
-    setStatusPopup(message);
-    setShowConcedeConfirm(false);
-    setActiveChallenge(null);
-  };
-
-  const endChallenge = (nextTurn: Player) => {
-    setTurnPlayer(nextTurn);
-    setPhase("idle");
-    setPickSeconds(PICK_SECONDS);
-    setQuestionSeconds(ANSWER_SECONDS);
-    setSwitchSeconds(ANSWER_SECONDS);
-    setActiveChallenge(null);
-  };
-
-  const loadQuestion = async (
-    row: number,
-    col: number,
-    player: Player,
-    questionType: QuestionType,
-  ) => {
-    if (gameState !== "playing") {
-      return;
-    }
-
-    const question = await getQuestion(questionType);
-
-    if (!question) {
-      updateTile(row, col, "black");
-      endChallenge(otherPlayer(player));
-      return;
-    }
-
-    setActiveChallenge({
-      row,
-      col,
-      ownerPlayer: player,
-      answerPlayer: player,
-      questionType,
-      question,
-    });
-    setQuestionSeconds(ANSWER_SECONDS);
-    setSwitchSeconds(ANSWER_SECONDS);
-    setPhase("answering");
-  };
-
-  const openStealChallenge = () => {
-    if (!activeChallenge || gameState !== "playing") {
-      return;
-    }
-
-    const stealPlayer = otherPlayer(activeChallenge.answerPlayer);
-    setActiveChallenge({
-      ...activeChallenge,
-      answerPlayer: stealPlayer,
-    });
-    setTurnPlayer(stealPlayer);
-    setQuestionSeconds(switchSeconds);
-    setStatusPopup(`${PLAYER_META[stealPlayer].label} accepted steal.`);
-    setPhase("stealing");
-  };
-
-  const revealAndContinue = (correctAnswer: string | boolean) => {
-    if (!activeChallenge || gameState !== "playing") {
-      return;
-    }
-
-    const { row, col, ownerPlayer } = activeChallenge;
-    const nextTurn = otherPlayer(ownerPlayer);
-
-    updateTile(row, col, "black");
-    setPhase("revealing");
-    setActiveChallenge(null);
-    setStatusPopup(`Correct answer: ${String(correctAnswer)}`);
-
-    window.setTimeout(() => {
-      setStatusPopup("");
-      setTurnPlayer(nextTurn);
-      setPhase("idle");
-      setPickSeconds(PICK_SECONDS);
-      setQuestionSeconds(switchSeconds);
-      setSwitchSeconds(switchSeconds);
-    }, ANSWER_REVEAL_MS);
-  };
-
-  const revealCorrectAndContinue = async () => {
-    if (!activeChallenge || gameState !== "playing") {
-      return;
-    }
-
-    const result = await checkQuestion({
-      question_id: activeChallenge.question.id,
-      answer: "",
-      question_type: activeChallenge.questionType,
-    });
-
-    revealAndContinue(result?.correct_answer ?? "n/a");
-  };
-
-  const handleAnswer = async (answer: string | boolean) => {
-    if (
-      !activeChallenge ||
-      gameState !== "playing" ||
-      (phase !== "answering" && phase !== "stealing")
-    ) {
-      return;
-    }
-
-    const response = await checkQuestion({
-      question_id: activeChallenge.question.id,
-      answer,
-      question_type: activeChallenge.questionType,
-    });
-
-    if (response?.is_correct) {
-      claimTile(
-        activeChallenge.row,
-        activeChallenge.col,
-        activeChallenge.answerPlayer,
-      );
-      setStatusPopup("");
-      endChallenge(otherPlayer(activeChallenge.answerPlayer));
-      return;
-    }
-
-    if (activeChallenge.questionType === "yes_no") {
-      const winnerPlayer = otherPlayer(activeChallenge.answerPlayer);
-      claimTile(activeChallenge.row, activeChallenge.col, winnerPlayer);
-      endChallenge(winnerPlayer);
-      return;
-    }
-
-    if (phase === "answering") {
-      setSwitchSeconds(ANSWER_SECONDS);
-      setTurnPlayer(otherPlayer(activeChallenge.answerPlayer));
-      setPhase("switch");
-      return;
-    }
-
-    revealAndContinue(response?.correct_answer ?? "n/a");
-  };
-
-  const handleTimeout = async () => {
-    if (!activeChallenge || gameState !== "playing") {
-      return;
-    }
-
-    if (phase === "answering" && activeChallenge.questionType === "standard") {
-      setSwitchSeconds(ANSWER_SECONDS);
-      setTurnPlayer(otherPlayer(activeChallenge.answerPlayer));
-      setPhase("switch");
-      return;
-    }
-
-    if (activeChallenge.questionType === "yes_no") {
-      const winnerPlayer = otherPlayer(activeChallenge.answerPlayer);
-      claimTile(activeChallenge.row, activeChallenge.col, winnerPlayer);
-      endChallenge(winnerPlayer);
-      return;
-    }
-
-    await revealCorrectAndContinue();
-  };
+  const {
+    board,
+    turnPlayer,
+    winner,
+    gameState,
+    gameResult,
+    phase,
+    pickSeconds,
+    questionSeconds,
+    switchSeconds,
+    statusPopup,
+    activeChallenge,
+    startNewGame,
+    startSelection,
+    handleAnswer,
+    openStealChallenge,
+    revealCorrectAndContinue,
+    concedeCurrentTurn,
+  } = usePyramidGameController();
 
   const pickPopupText =
     phase === "idle"
@@ -708,11 +179,8 @@ export default function PyramidLocalGame() {
         playerLabel={PLAYER_META[turnPlayer].label}
         onCancel={() => setShowConcedeConfirm(false)}
         onConfirm={() => {
-          const winnerPlayer = otherPlayer(turnPlayer);
-          endGame(
-            winnerPlayer,
-            `${PLAYER_META[turnPlayer].label} conceded. ${PLAYER_META[winnerPlayer].label} wins.`,
-          );
+          concedeCurrentTurn();
+          setShowConcedeConfirm(false);
         }}
       />
